+++ packages/core/CHANGELOG.md
@@ -1270,7 +1270,7 @@
 
 -   [#18](https://github.com/worktile/plait/pull/18) [`8c85cec`](https://github.com/worktile/plait/commit/8c85cec206736adcc57b9e0ead7cb14fc5ff17fe) Thanks [@Maple13](https://github.com/Maple13)! - feat(board): support readonly properties
 
-*   [`f76a4b1`](https://github.com/worktile/plait/commit/f76a4b15c884e1584bd52a7422d9b6134b3c9f7e) Thanks [@pubuzhixing8](https://github.com/pubuzhixing8)! - [core] fix listener keydown/keyup when board is selection mode
+*   [`f76a4b1`](https://github.com/worktile/plait/commit/f76a4b15c884e1584bd52a7422d9b6134b3c9f7e) Thanks [@pubuzhixing8](https://github.com/pubuzhixing8)! - [core] fix listener keydown/keyUp when board is selection mode
 
 ## 0.0.4